Add dependency injection example
This includes an example of a dependency injection to do testing with mock classes. A rudimentary database is provided via an interface (Database.kt) and its implementation (RealDatabase.kt). The database is stored as a reference in the CoachMeApplication class, so that it is available in all activities.

For testing, another Application class is created, CoachMeTestApplication, and it is instantiated by the MockTestRunner before launching the tests. CoachMeTestApplication overrides the database reference and replaces it with a reference to a mock database (MockDatabase).

This can be extended to do more dependency injection in testing by adding other dependencies in the CoachMeApplication class and replace them with their mock equivalents in CoachMeTestApplication class.