Add ShownEvent data class, implement basic multi day event wrapping
The way I solved the problem is that events are in backend/DB in form of Event instance, but everything in the frontend works events as ShownEvent instance. This instance additionally has attributes startText and endText to show the correct times for multi day events. The multiDayEventMap keeps track of multi day events (event -> List<ShownEvent>) for now.