Add code for maps testing (but with tests commented out) and small bug fix
After trying without success to test the map markers, I decided to keep the code that was added to test them (tags and callbacks), but the tests are commented out. That way, if we ever want to test it, we don't have to start from scratch, the tags are already here.